Refactored & updated README.
