+++ Sources/FaviconFinder/Finders/MockFaviconFinder.swift
@@ -45,7 +45,19 @@ class MockFaviconFinder: FaviconFinderProtocol {
                 source: URL(string: "https://google.com")!,
                 format: .appleTouchIcon,
                 sourceType: .html,
-                sizeTag: nil
+                sizeTag: "100x140"
+            ),
+            .init(
+                source: URL(string: "https://apple.com")!,
+                format: .appleTouchIcon,
+                sourceType: .html,
+                sizeTag: "100x90"
+            ),
+            .init(
+                source: URL(string: "https://facebook.com")!,
+                format: .appleTouchIcon,
+                sourceType: .html,
+                sizeTag: "100x90"
             )
         ]
     }
